feat(mcp): validate .loopover.yml offline by extracting buildFocusManifestValidation into @loopover/engine
loopover_validate_config's result builder, buildFocusManifestValidation, lived app-only in
src/services/focus-manifest-validation.ts, so the local stdio server had to POST to
/v1/validate/focus-manifest to validate a manifest -- unusable offline. Its parsing core
(parseFocusManifestContent) was already engine-side, but the result builder and its
unknownTopLevelWarnings dependency (src/selfhost/config-lint.ts) were not.

Extract both into @loopover/engine (packages/loopover-engine/src/{focus-manifest-validation,config-lint}.ts),
leaving re-export shims at the original src/ locations so the remote server (src/mcp/server.ts) and
API (src/api/routes.ts) keep working unchanged. Add the two exports-map entries and barrel re-exports.
Rewire the local loopover_validate_config handler to call buildFocusManifestValidation in-process
instead of apiPost, so a user validates a .loopover.yml fully offline.

All existing focus-manifest-validation and config-lint tests pass against the new location; adds a
stdio test that drives loopover_validate_config with an unreachable API URL to prove offline behavior.

Closes #6269
